requirements.txt
discord.py==2.4.0
docker==7.1.0